leveldb/storage: catch EINVAL on go >= 1.8 too
In go 1.8, various functions were updated to return PathError instead.
This includes `os.File Sync()`.

For reference, the upstream commit which made this change is
https://github.com/golang/go/commit/212d2f82e05.

Arguably we could take this as a signal that upstream doesn't consider
this error type stable, so rather than attempt to identify it we could
assume any error from Sync should be ignored, but I think there's some
amount of value still in trying to error out on things like ENOSPACE or
EBADF.